requirements.txt
Pillow==10.4.0
pillow-heif==0.16.0
python-magic==0.4.27
